fix: crypto is not accessible in a Worker scope
