+++ cli/tests/039_worker_deno_ns/has_ns.ts
@@ -2,7 +2,7 @@ onmessage = (msg): void => {
   if (msg.data === "START") {
     postMessage("has_ns.ts: is window.Deno available: " + !!window.Deno);
   } else {
-    const worker = new Worker("./maybe_ns.ts");
+    const worker = new Worker("./maybe_ns.ts", { type: "module", deno: true });
     worker.onmessage = (msg): void => {
       postMessage("[SPAWNED BY has_ns.ts] " + msg.data);
     };
