+++ cli/tests/039_worker_deno_ns/no_ns.ts
@@ -2,7 +2,7 @@ onmessage = (msg): void => {
   if (msg.data === "START") {
     postMessage("no_ns.ts: is window.Deno available: " + !!window.Deno);
   } else {
-    const worker = new Worker("./maybe_ns.ts");
+    const worker = new Worker("./maybe_ns.ts", { type: "module" });
     worker.onmessage = (msg): void => {
       postMessage("[SPAWNED BY no_ns.ts] " + msg.data);
     };
