+++ cli/tests/040_worker_blob.ts
@@ -1,6 +1,6 @@
 const b = new Blob(["console.log('code from Blob'); postMessage('DONE')"]);
 const blobURL = URL.createObjectURL(b);
-const worker = new Worker(blobURL);
+const worker = new Worker(blobURL, { type: "module" });
 worker.onmessage = (): void => {
   Deno.exit(0);
 };
