+++ cli/tests/integration_tests.rs
@@ -1080,8 +1080,7 @@ itest!(_038_checkjs {
   output: "038_checkjs.js.out",
 });
 
-// TODO(bartlomieju): re-enable
-itest_ignore!(_039_worker_deno_ns {
+itest!(_039_worker_deno_ns {
   args: "run --reload 039_worker_deno_ns.ts",
   output: "039_worker_deno_ns.ts.out",
 });
